requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1
django-allauth==64.1.0
django-crispy-forms==2.3
django-environ==0.11.2
django-import-export==4.1.1
django-jazzmin==3.0.0
django-mathfilters==1.0.0
django-simple-history==3.7.0
djangorestframework==3.15.2
dnspython==2.6.1
mysqlclient==2.2.4
pillow==10.4.0
pymongo==4.8.0
python-decouple==3.8
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
whitenoise==6.7.0
